support TOML configuration files
This is just a stub, but good to get some feedback on. I decided to wrap
the Dict returned by the TOML parser into a Config struct. This way, we
can easily define functions that operate on the configuration.
Specifically, I already implemented Base.getproperty, to make it easy to
get values with dot notation:

config.input.time instead of config["input"]["time"]