restructure config; add name mapping
The mapping in name.jl is not used yet in this commit, it is only added.

This also adds a CSV option to the Writer, which is not yet active. First we will refactor the code a bit to use the new parameter mapping, and to fully switch to using structs of arrays, not arrays of structs.