remove TypedTables and FillArrays dependencies
TypedTables is no longer needed since we are switching to structs of vectors.

FillArrays could still save some memory in some cases, but then we have to take into account that we cannot update the values in that array like we can for all other parameters. This extra complexity is probably not worth the reduced memory usage.