reservoir and lake from array to structs to struct of arrays
Currently the length of the arrays is equal to the length of river cells. An extra entry `is_lake/is_res` was added to indicate if it is an active lake or reservoir. The next step will be to use vectors with length equal to the amount of reservoirs or lakes, which will save a lot of memory. To prevent the diff becoming too large, we pause here.