store only actual reservoirs and lakes
Not one entry for each river that is only valid if there is an actual reservoir or lake there.

This also adds some inner constructors to guarantee that all vectors passed to our structs of arrays are of equal length, to catch possible mistakes. (Such as `wb_pits`, which is fixed in this commit)