include config in the model object
Some of the information in the config is already parsed and processed into other components of the model object. However it is convenient to hold on to the config object itself as well, for instance for settings that are needed after initializing the model. It is not desirable to have to add these one by one to the model struct.