refactor: add toposort to network
Before this the network in the model was a namedtuple with keys `land` and `river` and their graphs as values. Now instead each domain (land/river) has it's own namedtuple, which stores both the graph and the topological sort together.

This will simplify the code and reduce the amount of arguments that have to be passed around when running the code.