CSV support selections, hook up paramap
Two distinct changes that became intertwined.
CSV now supports 4 new ways to select which cells you want to report on. The default is all active cells if nothing is specified. We already supported `index`, but this is only useful if you know the internal vector index. The 4 new ways are:

1. coordinate: specify an x and y coordinate, and the nearest cell will be selected
2. index: specify an x and y index, and that cell will be selected
3. pointmap: used to output gauges, i.e. unique cells from a map
4. areamap: used to reduce areas such as subcatchments

The paramap is now used when constructing the model, as well as updating the forcings. Some renamings were needed, as I'd like the parameter key names in the TOML to be identical to the key in paramap. A new `key` keyword argument was added to `ncread` to convert the internal name to the external one.