remove internal parameter map
This updates the config.toml such that the names and locations directly correspond to the structs and their location in the model struct, removing a level of indirection, and making it easier to extend the code without keeping an internal directory (paramap).

This should allow but does not yet include refactoring and simplification of the initialization code.

Fixes #36