drop DataFrames and CSV dependencies
This speeds up loading a precompiled Wflow from 4.3 to 3.2 seconds on my machine, julia 1.6.

We weren't really using many of the features of these two packages, so it was relatively easy to take out. DelimitedFiles is much less capable than CSV, but is a Stdlib and suffices for us. I wrote a helper function to have it spit out a NamedTuple like CSV does.