add ProgressLogging to simulation loop
This makes it run the simulation loop over the range starttime:dt:endtime. This is slightly different than the `is_finished` used before, hence two test values need to be updated.

Besides the progressmeter basic model information is shown before the run. We still suppress all logging output for testing.

Note that by default the progress bar is not shown. This only happens in VS Code / Juno or when using the TerminalLoggers logger. We will use the logger in the wflow_cli app.