close the staticmaps NetCDF during initialization
If the cyclic data is in there as well, the NCReader also opens the same file, but that is a different handle that is already being closed properly. With this, all files are closed, tested by running a simulation, and before closing the session, moving all input+output data to a different folder. That's how I found this file was left open as well.