make state optional
Input state was quite easy, just read them if reinit is false.
For output state we now allow the state Dataset in the Writer to be nothing.
And we also allow the whole state section to be absent.
Test this by running a short run of a minimal sbm model.