use a single internal dimension order with increasing axes
Previously we accepted x and y dimensions in either order, and adapted to that internally. However, some issues arose with other variants present in and between the netCDF inputs that made it clear that we need to support more variants. Some benchmarking showed that moving away from in place netCDF loading hardly affected performance. So to be both more flexible, and reduce some internal code complexity, we treat the data at entry to always follow the dimension order (x, y, layer, time), and to have increasing axes for each dimension.