use CairoMakie to make the plots
The Plots figures also work, but I plan to add more model visualization based on Makie as well, so best to keep it to a single package.

This will embed a PNG in the HTML webpage, so no need to save the file. Downside is that we have to use example blocks with hide comments to not print the code.