Function `save_state` (#370)
* Close state file in `save_state` function

* Check if state dataset has been closed in `finalize`
In the `save_state` function the state dataset is closed after saving the state variables to a NetCDF file. When this function is called before `finalize` a check is required.

* Fix check in `finalize`

* `ShallowWaterRiver` lake and reservoir index
Do not exchange (BMI) these variables. The size is not equal to the `river` grid size, but to the `lake` or `reservoir` size.

* Add test closing of states dataset

* Update changelog