Remove lower limit of zero for `sbm.runoff` (#373)
* Remove lower limit of zero for `sbm.runoff`
Setting a lower limit of zero for `sbm.runoff` originates from the wflow_sbm Python implementation, when surface runoff was computed with one kinematic wave, to avoid a very slow kinematic wave solution. In the Python code the kinematic wave was split into a river and land component. While for river runoff the lower limit was removed, this was not done for land runoff yet.

* Change description `runoff`

* Extend `sbm` tests
Compute mean values of variables runoff, soilevap and snow, and add variables actevap and actinfilt.

* Update changelog

* Add `net_runoff` variable to `SBM`

* Fix `WflowServer` tests

* Update docs and changelog

* Update description `runoff` variable `SBM` struct

Co-authored-by: JoostBuitink <44062204+JoostBuitink@users.noreply.github.com>

---------

Co-authored-by: Carlos Fernando Baptista <cfd.baptista@gmail.com>
Co-authored-by: JoostBuitink <44062204+JoostBuitink@users.noreply.github.com>