Change use of routing water depth `h_av` to `h`  (#555)
The routing average water depth `h_av` at the previous model time step is used to limit open water evaporation and to set the river `stage` of the `GroundwaterFlow` model for the current model time step. It is more accurate to use the instantaneous value `h` from the previous model time step.