Edit "Installing as Julia package"
The current link https://julialang.org/downloads/#current_stable_release goes to the old way to install. As you can see on top of that page the current recommenden way is with juliaup, which is described here: https://julialang.org/install/.

It is not the official docs, but for new users I think https://modernjuliaworkflows.org/writing/#installation is more accessible, so I link there instead.