Rootfraction unsaturated zone (#611)
* Root fraction unsaturated zone
The root fraction (per soil layer) in the unsaturated zone is now scaled by the sum of the rootfraction in the unsaturated zone. Plant transpiration can then be fully satisfied from the unsaturated zone (depending on the water table depth). For the transpiration from the saturated store the remaining potential transpiration is used in combination with the wet root fraction (rootfraction in the saturated soil layers is not used).

* Update wflow tests

* Update wflow server tests

* Update changelog

* Update docs

* Remove  `rootfraction_unsat` from `SbmSoilVariables`
Simplified the computation a bit for sum of root fraction unsaturated soil layers and scaling of unsaturated root fraction.

* fix init floodplain storage

* get width and length from domain.river

* Remove `river =` from init floodplain storage

---------

Co-authored-by: JoostBuitink <44062204+JoostBuitink@users.noreply.github.com>