Support use of observed reservoir outflow (#683)
* Add reservoir variable `outflow_obs`

* Add reservoir function for `outflow_obs`

* Remove two reservoir variables
As the variables `demandrelease` and `percfull` are saved at the internal timestep and are not part of model output (standard names). it is not that useful to store these as  part of the simple reservoir model.

* Add standard name
For observed reservoir outflow.

* Add unit test reservoir observed outflow

* Add test with cyclic observed reservoir outflow
For Piave test model.

* Add comment and use `MISSING_VALUE`
instead of `NaN`.

* Update reservoir.qmd

* Update parameters_routing.qmd

* Update changelog

* Remove `print` statements

* Add reservoir location id to parameters

* Add `debug` message
For using reservoir observed outflow. This is done at the model timestep level.

* Add test debug log message
For using observed reservoir outflow.

* Address review comment
Throw log debug message for using observed outflow at reservoir locations when input parameter "reservoir_water~outgoing~observed__volume_flow_rate" is defined.