Improve accuracy solving kinematic wave subsurface flow (#766)
* Improve accuracy kinematic wave subsurface flow solution
This is for subsurface flow with `KhExponential` and `KhExponentialConstant` depth profiles. The previous solution is a bit hard to follow and quite similar to using an average celerity for the water table depth change (kinematic wave celerity is a function of depth). Using a maximum allowed water table depth change per iteration is more accurate based on a couple of tests and implemented for these two saturated hydraulic conductivity depth profiles.

* Fix exfiltration rate computation

* Update tests

* Rename couple of internal variables

* Change unit recharge subsurface flow

* Some small improvements

* Add docstrings and comments

* Remove unused functions

* Fix kinematice wave `celerity` profile `KhExponentialConstant`

* Add unit test
For kinematic wave subsurface flow with `KhExponentialConstant` profile.

* Update changelog

* Fix equation docs
Computation of celerity for `exponential` and `exponential_constant` profiles kinematic wave subsurface flow.

* Address review comment